Move SSE logic from AlertsPanel to EventsPanel and improve confidence display
SSE Refactoring:
- Move all SSE connection management from AlertsPanel to EventsPanel
- EventsPanel now owns both SSE connection and event display
- AlertsPanel simplified to only handle Alerts API
- Remove cross-component bridging through Home.vue for SSE events

Confidence Display:
- Show count of confidence values when multiple: "47%-65% confidence (2)"
- Camera info panel shows: "Confidence: between 40% and 91% (3)"
