+++ src/components/AlertsPanel.vue
@@ -1,15 +1,12 @@
 <script setup lang="ts">
 import { ref, watch, computed, onMounted, onUnmounted } from 'vue'
 import {
-  createEventSubscription,
-  connectToEventSubscription,
-  deleteEventSubscription,
   listAlerts,
   listNotifications,
   listEventAlertConditionRules,
   listAlertActions
 } from 'een-api-toolkit'
-import type { Camera, EenError, SSEEvent, SSEConnection, SSEConnectionStatus, Alert, Notification, EventAlertConditionRule, AutomationAlertAction } from 'een-api-toolkit'
+import type { Camera, EenError, Alert, Notification, EventAlertConditionRule, AutomationAlertAction } from 'een-api-toolkit'
 
 // Extended alert type with optional notification and rule data
 interface AlertWithNotification extends Alert {
@@ -26,16 +23,13 @@ const props = defineProps<{
   activeAlertId?: string | null
   initialDuration?: number | null
   initialAutoRefresh?: boolean
-  initialLiveFeed?: boolean
   initialEventFilter?: boolean
 }>()
 
 const emit = defineEmits<{
-  (e: 'sse-event', event: Record<string, unknown>): void
   (e: 'alert-clicked', alert: { alertId: string; alertObject: Record<string, unknown> }): void
   (e: 'duration-changed', duration: number): void
   (e: 'auto-refresh-changed', enabled: boolean): void
-  (e: 'live-feed-changed', enabled: boolean): void
   (e: 'event-filter-changed', enabled: boolean): void
 }>()
 
@@ -44,13 +38,6 @@ const { loadImage, getImage } = useImageCache()
 
 // Use event age formatting (updates every second)
 const { formatAge } = useEventAge()
-
-// SSE State
-const subscriptionId = ref<string | null>(null)
-const sseConnection = ref<SSEConnection | null>(null)
-const connectionStatus = ref<SSEConnectionStatus>('disconnected')
-const connectionError = ref<EenError | null>(null)
-const liveFeedEnabled = ref(props.initialLiveFeed ?? false) // User's desired state - always reconnect when enabled
 
 // Alerts State
 const alerts = ref<AlertWithNotification[]>([])
@@ -111,25 +98,6 @@ const getInitialDuration = () => {
 }
 const selectedTimeRange = ref(getInitialDuration())
 
-// Reconnection timer (SSE subscriptions expire after 15 minutes)
-const SUBSCRIPTION_TTL_MS = 14 * 60 * 1000 // Reconnect at 14 minutes (before 15 min expiry)
-let reconnectTimer: ReturnType<typeof setTimeout> | null = null
-let isProactiveReconnecting = false // Flag to prevent double-reconnection
-
-// Debounce timer for event type changes
-const DEBOUNCE_DELAY_MS = 500 // Wait 500ms for event type changes to settle
-let debounceTimer: ReturnType<typeof setTimeout> | null = null
-
-// Connection guard - tracks current connection attempt to prevent races
-let connectionAttemptId = 0
-
-// Computed
-const isConnected = computed(() => connectionStatus.value === 'connected')
-const isConnecting = computed(() => connectionStatus.value === 'connecting')
-const canConnect = computed(() => {
-  return props.camera && props.selectedTypes.length > 0 && !isConnected.value && !isConnecting.value
-})
-
 // Alerts computed
 const hasMoreAlerts = computed(() => !!alertsNextPageToken.value)
 const hasNoAlerts = computed(() => !alertsLoading.value && alerts.value.length === 0 && !alertsError.value)
@@ -158,41 +126,6 @@ const eventFilterButtonClass = computed(() => {
   return 'bg-gray-500 hover:bg-gray-600 text-white'
 })
 
-// Computed button label based on state
-const feedButtonLabel = computed(() => {
-  if (isConnecting.value) return 'Connecting...'
-  if (isConnected.value) return 'Disable Live Events'
-  return 'Turn Live Events On'
-})
-
-// Computed button styling based on state
-const feedButtonClass = computed(() => {
-  if (isConnecting.value) {
-    return 'bg-yellow-600 hover:bg-yellow-700 text-white'
-  }
-  if (isConnected.value) {
-    return 'bg-green-600 hover:bg-green-700 text-white'
-  }
-  return 'bg-gray-500 hover:bg-gray-600 text-white'
-})
-
-// Toggle live feed on/off
-function toggleLiveFeed() {
-  if (isConnected.value || isConnecting.value) {
-    // Turn off - disconnect and disable
-    liveFeedEnabled.value = false
-    emit('live-feed-changed', false)
-    disconnect()
-  } else {
-    // Turn on - enable and connect
-    liveFeedEnabled.value = true
-    emit('live-feed-changed', true)
-    if (canConnect.value) {
-      connect()
-    }
-  }
-}
-
 // Toggle event type filter for alerts
 function toggleEventFilter() {
   eventFilterEnabled.value = !eventFilterEnabled.value
@@ -200,276 +133,6 @@ function toggleEventFilter() {
   // Refresh alerts with new filter setting
   fetchAlerts()
 }
-
-// Handle new SSE event - just emit to parent
-function handleEvent(event: SSEEvent) {
-  // Emit SSE event to parent for insertion into historic events
-  emit('sse-event', event as unknown as Record<string, unknown>)
-}
-
-// Clear the reconnect timer
-function clearReconnectTimer() {
-  if (reconnectTimer) {
-    clearTimeout(reconnectTimer)
-    reconnectTimer = null
-  }
-}
-
-// Clear the debounce timer
-function clearDebounceTimer() {
-  if (debounceTimer) {
-    clearTimeout(debounceTimer)
-    debounceTimer = null
-  }
-}
-
-// Start the reconnect timer (proactive reconnection before TTL expires)
-function startReconnectTimer() {
-  clearReconnectTimer()
-
-  if (!liveFeedEnabled.value) {
-    return
-  }
-
-  reconnectTimer = setTimeout(() => {
-    if (liveFeedEnabled.value && props.camera && props.selectedTypes.length > 0) {
-      // Don't clear events on proactive reconnect
-      reconnect()
-    }
-  }, SUBSCRIPTION_TTL_MS)
-}
-
-// Reconnect without clearing events
-async function reconnect() {
-  // Set flag to prevent handleStatusChange from triggering another connect
-  isProactiveReconnecting = true
-
-  // Close existing connection
-  if (sseConnection.value) {
-    sseConnection.value.close()
-    sseConnection.value = null
-  }
-
-  // Delete old subscription
-  if (subscriptionId.value) {
-    await deleteEventSubscription(subscriptionId.value)
-    subscriptionId.value = null
-  }
-
-  // Create new subscription (don't clear events)
-  await connectWithoutClearingEvents()
-
-  // Clear the flag
-  isProactiveReconnecting = false
-}
-
-// Handle status change
-function handleStatusChange(status: SSEConnectionStatus) {
-  connectionStatus.value = status
-
-  if (status === 'error' || status === 'disconnected') {
-    clearReconnectTimer()
-
-    // Clean up if connection lost
-    if (sseConnection.value) {
-      sseConnection.value = null
-    }
-
-    // Skip auto-reconnect if we're doing a proactive reconnection
-    if (isProactiveReconnecting) {
-      return
-    }
-
-    // Auto-reconnect if feed is enabled and we have camera/types
-    if (liveFeedEnabled.value && props.camera && props.selectedTypes.length > 0) {
-      // Small delay before reconnecting
-      setTimeout(() => {
-        if (liveFeedEnabled.value && !isConnected.value && !isConnecting.value && !isProactiveReconnecting) {
-          connect()
-        }
-      }, 2000)
-    }
-  } else if (status === 'connected') {
-    // Start the proactive reconnect timer when connected
-    startReconnectTimer()
-  }
-}
-
-// Handle SSE error
-function handleError(error: Error) {
-  connectionError.value = {
-    code: 'NETWORK_ERROR',
-    message: error.message || 'SSE connection error'
-  }
-}
-
-// Connect to SSE stream
-async function connect() {
-  if (!props.camera || props.selectedTypes.length === 0) return
-
-  // Increment connection attempt ID to invalidate any pending connections
-  connectionAttemptId++
-  const currentAttemptId = connectionAttemptId
-
-  connectionStatus.value = 'connecting'
-  connectionError.value = null
-
-  await createAndConnectSubscription(currentAttemptId)
-}
-
-// Connect without clearing events (for proactive reconnection)
-async function connectWithoutClearingEvents() {
-  if (!props.camera || props.selectedTypes.length === 0) return
-
-  // Increment connection attempt ID to invalidate any pending connections
-  connectionAttemptId++
-  const currentAttemptId = connectionAttemptId
-
-  connectionStatus.value = 'connecting'
-  connectionError.value = null
-
-  await createAndConnectSubscription(currentAttemptId)
-}
-
-// Core subscription creation and connection logic
-async function createAndConnectSubscription(attemptId: number) {
-  // Create subscription
-  const subscriptionResult = await createEventSubscription({
-    deliveryConfig: { type: 'serverSentEvents.v1' },
-    filters: [{
-      actors: [`camera:${props.camera!.id}`],
-      types: props.selectedTypes.map(type => ({ id: type }))
-    }]
-  })
-
-  // Check if this attempt is still current (another connection might have been requested)
-  if (attemptId !== connectionAttemptId) {
-    // This attempt is stale, clean up and abort
-    if (!subscriptionResult.error && subscriptionResult.data?.id) {
-      deleteEventSubscription(subscriptionResult.data.id)
-    }
-    return
-  }
-
-  if (subscriptionResult.error) {
-    connectionError.value = subscriptionResult.error
-    connectionStatus.value = 'error'
-    return
-  }
-
-  subscriptionId.value = subscriptionResult.data.id
-
-  // Get SSE URL
-  const sseUrl = subscriptionResult.data.deliveryConfig.type === 'serverSentEvents.v1'
-    ? subscriptionResult.data.deliveryConfig.sseUrl
-    : null
-
-  if (!sseUrl) {
-    connectionError.value = {
-      code: 'API_ERROR',
-      message: 'No SSE URL returned from subscription'
-    }
-    connectionStatus.value = 'error'
-    return
-  }
-
-  // Check again before connecting
-  if (attemptId !== connectionAttemptId) {
-    // This attempt is stale, clean up and abort
-    deleteEventSubscription(subscriptionResult.data.id)
-    subscriptionId.value = null
-    return
-  }
-
-  // Connect to SSE stream
-  const connectionResult = connectToEventSubscription(sseUrl, {
-    onEvent: handleEvent,
-    onError: handleError,
-    onStatusChange: handleStatusChange
-  })
-
-  if (connectionResult.error) {
-    connectionError.value = connectionResult.error
-    connectionStatus.value = 'error'
-    return
-  }
-
-  sseConnection.value = connectionResult.data
-}
-
-// Disconnect from SSE stream
-async function disconnect() {
-  // Increment attempt ID to cancel any pending connection attempts
-  connectionAttemptId++
-
-  // Clear timers
-  clearReconnectTimer()
-  clearDebounceTimer()
-
-  // Close SSE connection
-  if (sseConnection.value) {
-    sseConnection.value.close()
-    sseConnection.value = null
-  }
-
-  // Delete subscription
-  if (subscriptionId.value) {
-    await deleteEventSubscription(subscriptionId.value)
-    subscriptionId.value = null
-  }
-
-  connectionStatus.value = 'disconnected'
-  connectionError.value = null
-}
-
-// Debounced reconnection handler - only reconnects if liveFeedEnabled
-async function debouncedReconnect(cameraId: string, types: string[]) {
-  // Clear any existing debounce timer
-  clearDebounceTimer()
-
-  // First, disconnect any existing connection immediately
-  const wasConnected = isConnected.value || isConnecting.value
-  if (wasConnected) {
-    await disconnect()
-  }
-
-  // If no types selected or feed not enabled, don't reconnect
-  if (types.length === 0 || !liveFeedEnabled.value) {
-    return
-  }
-
-  // Debounce the reconnection to wait for rapid changes to settle
-  debounceTimer = setTimeout(async () => {
-    debounceTimer = null
-    // Double-check we still have valid camera, types, and feed is enabled
-    if (props.camera?.id === cameraId && props.selectedTypes.length > 0 && liveFeedEnabled.value) {
-      await connect()
-    }
-  }, DEBOUNCE_DELAY_MS)
-}
-
-// Watch for camera or selected types changes - reconnect automatically with debouncing
-watch(
-  [() => props.camera?.id, () => props.selectedTypes],
-  ([newCameraId, newTypes], [oldCameraId, oldTypes]) => {
-    const cameraChanged = newCameraId !== oldCameraId
-    const typesChanged = JSON.stringify(newTypes) !== JSON.stringify(oldTypes)
-
-    // Only act if something actually changed
-    if (!cameraChanged && !typesChanged) {
-      return
-    }
-
-    // Use debounced reconnection
-    if (newCameraId && newTypes) {
-      debouncedReconnect(newCameraId, newTypes)
-    } else if (!newCameraId || newTypes.length === 0) {
-      // No camera or no types - just disconnect
-      disconnect()
-    }
-  },
-  { deep: true }
-)
 
 // ==================== ALERTS FUNCTIONALITY ====================
 
@@ -954,33 +617,17 @@ if (props.camera) {
   fetchAlerts()
 }
 
-// Start auto-refresh timer and live feed on mount if enabled
+// Start auto-refresh timer on mount if enabled
 onMounted(() => {
   if (autoRefresh.value) {
     startRefreshTimer()
   }
-  // Start live feed connection if enabled and we have camera and types
-  if (liveFeedEnabled.value && props.camera && props.selectedTypes.length > 0) {
-    connect()
-  }
 })
 
 // Cleanup on unmount
-onUnmounted(async () => {
-  await disconnect()
+onUnmounted(() => {
   stopRefreshTimer()
   document.removeEventListener('keydown', handleModalEscKey)
-})
-
-// Expose toggle function and state for parent component
-defineExpose({
-  toggleLiveFeed,
-  feedButtonLabel,
-  feedButtonClass,
-  canConnect,
-  isConnected,
-  isConnecting,
-  connectionError
 })
 </script>
